Latest solution. AC on 25th January 2019.
Used BST iterator. The approach is use the logic for 2 Sum in a sorted array. Create one BST iterator which points to the smallest element and create another which points to largest element. Use next() method to determine if sum is equal, greater or less than expected and move forward accordingly.